\name{HSLuv-package}
\alias{HSLuv-package}
\alias{HSLuv}
\docType{package}
\title{
  Converts HSLuv to RGB and Hex
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Scott Spencer, spencer.sa@gmail.com

Maintainer: Scott Spencer <spencer.sa@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
