\name{hsluv_rgb}
\alias{hsluv_rgb}
\docType{package}
\title{
hsluv_rgb converts HSLuv values to RGB values
}
\description{
hsluv_rgb converts HSLuv values to RGB values.
}
\usage{
hsluv_rgb(hsl)	
}
\arguments{
\item{hsl}{Hue (0, 360), saturation (0, 100), and luminance (0, 100) as a vector of integers.}
}
\examples{
\dontrun{
hsluv_rgb(c(2, 2, 2))
}
}
